Change how resultsets handle column names
Resultsets now track names of their columns.  They can have names added
to columns with add_column_name() (used by CsvSelect::File).  When resultsets
are joined, the new resultset inherits the column names of the resultsets
that went into constructing it.

This allows specifying columns by name in joins and shows without having to
track the column indexes as new files are joined to the right